# docs/swagger.yaml
definitions:
  Dto.Avatar:
    properties:
      color:
        description: hex color code
        type: string
      hash:
        type: string
      kind:
        description: e.g., "user", "org"
        type: string
      name:
        type: string
    type: object
...
  Dto.Repository:
    properties:
      description:
        type: string
      is_public:
        type: boolean
      is_starred:
        type: boolean
      kind:
        description: Image or Application
        type: string
      name:
        description: Image name
        type: string
      namespace:
        type: string
      state:
        type: string
    type: object
...
  Dto.RepositoryPermission:
    properties:
      avatar:
        $ref: '#/definitions/Dto.Avatar'
      isRobot:
        description: Optional fields
        type: boolean
      name:
        type: string
      role:
        type: string
    type: object
...
  Dto.UpdateRepositoryPermission:
    properties:
      role:
        description: 'allowed value: "admin", "write", "read"'
        type: string
    type: object
...
  Errors.ErrorDetails:
    properties:
      code:
        description: Code is a sting that represents the error code (e.g. "USER_NOT_FOUND",
          "INVALID_INPUT", etc.)
        type: string
      message:
        type: string
    type: object
  Errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/Errors.ErrorDetails'
    type: object
info:
  contact: {}
  description: Quay registry API implemented in Go
  title: Quay Go API
  version: "1.0"
paths:
...
  /api/v1/repository/{repository}/permissions/team:
    get:
      description: List teams permission on a repository
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Dto.RepositoryPermission'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List teams permission on a repository
      tags:
      - Permission
  /api/v1/repository/{repository}/permissions/team/{teamname}:
    get:
      description: Get a team's permission on a repository
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      - description: Team name
        in: path
        name: teamname
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.RepositoryPermission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a team's permission on a repository
      tags:
      - Permission
    patch:
      consumes:
      - application/json
      description: Update team repository permission
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      - description: Team name
        in: path
        name: teamname
        required: true
        type: string
      - description: Repository details to change
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/Dto.UpdateRepositoryPermission'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.Repository'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update team repository permission
      tags:
      - Permission
  /api/v1/repository/{repository}/permissions/user:
    get:
      description: List users permission on a repository
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Dto.RepositoryPermission'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List users permission on a repository
      tags:
      - Permission
  /api/v1/repository/{repository}/permissions/user/{username}:
    get:
      description: Get a user's permission on a repository
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      - description: Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.RepositoryPermission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a user's permission on a repository
      tags:
      - Permission
    patch:
      consumes:
      - application/json
      description: Update user repository permission
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Repository details to change
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/Dto.UpdateRepositoryPermission'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.Repository'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user repository permission
      tags:
      - Permission
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication using the Authorization header. The value
      should be in the format "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"